Add authorization to endpoints, CORS middleware and rate limiter

+++ app/main.py
@@ -1,14 +1,17 @@
 from datetime import datetime, timedelta, timezone
-
+import time
 import jwt
 from jwt.exceptions import InvalidTokenError
-from fastapi import FastAPI, HTTPException, Depends
+from fastapi import FastAPI, HTTPException, Depends, Request
 from fastapi.security import OAuth2PasswordBearer, OAuth2PasswordRequestForm
+from fastapi.responses import JSONResponse
 from sqlmodel import SQLModel, create_engine, Session, select, Field, Column, JSON
 from fastapi.middleware.cors import CORSMiddleware
 from typing import Optional, List, Annotated
-from pydantic import BaseModel
+from pydantic import BaseModel, EmailStr
 from pwdlib import PasswordHash
+from rate_limiter import RateLimiterStore
+from pathlib import Path
 
 import os
 from dotenv import load_dotenv
@@ -17,7 +20,7 @@ load_dotenv()
 
 secret_key = os.getenv("secret_key")
 algorithm = "HS256"
-access_token_minutes = 30
+access_token_minutes = 60
 
 class Token(BaseModel):
     access_token: str
@@ -28,20 +31,18 @@ class User(SQLModel, table = True):
 
     username: str = Field(unique = True)
     display_name: str
-    email: str
+    email: EmailStr = Field(unique = True)
     hashed_password: str
 
 class UserCreate(SQLModel):
-    username: str
+    username: str = Field(min_length = 3)
     display_name: str
-    email: str
-    password: str
+    email: EmailStr
+    password: str = Field(min_length = 8)
 
 class UserPublic(SQLModel):
-    id: int
     username: str
     display_name: str
-    email: str
 
 class ExperimentBase(SQLModel):
     title: str
@@ -71,6 +72,7 @@ class ExperimentUpdate(SQLModel):
 
 class Experiment(ExperimentBase, table = True):
     id: Optional[int] = Field(default = None, primary_key = True)
+    owner_id: int = Field(foreign_key = "user.id")
 
 class ExperimentCreate(ExperimentBase):
     pass
@@ -78,7 +80,8 @@ class ExperimentCreate(ExperimentBase):
 class ExperimentPublic(ExperimentBase):
     id: int
 
-sqlite_database = "database.db"
+Path("data").mkdir(exist_ok = True)
+sqlite_database = "data/database.db"
 sqlite_url = f"sqlite:///{sqlite_database}"
 
 connect_args = {"check_same_thread": False}
@@ -90,6 +93,40 @@ def create_db():
 password_hash = PasswordHash.recommended()
 
 app = FastAPI(title = "Hypothesis API", description = "An api for storing and retreiving scientific experiment logs")
+
+origins = [
+    "https://rasmusstenlund.github.io",
+    "https://hypothesis.dino.icu"
+]
+
+app.add_middleware(
+    CORSMiddleware,
+    allow_origins = origins,
+    allow_credentials=False,
+    allow_methods=["*"],
+    allow_headers=["*"],
+)
+
+limiter = RateLimiterStore(max_tokens = 60, refill_rate = 10, interval = 60)
+
+@app.middleware("http")
+async def rate_limit_middleware(request: Request, call_next):
+    client_ip = request.client.host
+    bucket = limiter.get_bucket(client_ip)
+
+    if not bucket.allow_request():
+        retry_after = bucket.get_reset_time() - time.time()
+        return JSONResponse(
+            status_code = 429,
+            content = {"detail": "Too many requests. Try again later."},
+            headers = {
+                "Retry-After": str(max(1, int(retry_after))),
+            }
+        )
+    
+    response = await call_next(request)
+
+    return response
 
 def create_access_token(data):
     to_encode = data.copy()
@@ -145,19 +182,33 @@ async def get_current_user(token: Annotated[str, Depends(oauth2_scheme)]):
 def on_startup():
     create_db()
 
+@app.get("/")
+def root():
+    return {"message": "Hypothesis API is running"}
+
+@app.get("/users/me", response_model = UserPublic)
+def get_me(current_user: User = Depends(get_current_user)):
+    return current_user
+
 @app.post("/users", response_model = UserPublic)
 def create_user(user: UserCreate):
     existing_user = get_user(user.username)
     if existing_user:
         raise HTTPException(
             status_code = 400,
-            detail = "Username already exists"
+            detail = "Username or email already in use"
         )
     hashed_password = get_password_hash(user.password)
 
     db_user = User(username = user.username, display_name = user.display_name, email = user.email, hashed_password = hashed_password)
 
     with Session(engine) as session:
+        existing_email = session.exec(select(User).where(User.email == user.email)).first()
+        if existing_email:
+            raise HTTPException(
+                status_code = 400,
+                detail = "Username or email already in use"
+            )
         session.add(db_user)
         session.commit()
         session.refresh(db_user)
@@ -183,7 +234,7 @@ async def login(form_data: Annotated[OAuth2PasswordRequestForm, Depends()]):
 @app.post("/experiments", response_model = ExperimentPublic)
 async def create_experiment(current_user: Annotated[User, Depends(get_current_user)], experiment: ExperimentCreate):
     with Session(engine) as session:
-        db_experiment = Experiment.model_validate(experiment)
+        db_experiment = Experiment(**experiment.model_dump(), owner_id =  current_user.id)
 
         session.add(db_experiment)
         session.commit()
@@ -198,6 +249,11 @@ async def update_experiment(current_user: Annotated[User, Depends(get_current_us
             raise HTTPException(
                 status_code = 404,
                 detail = "Experiment not found"
+            )
+        if db_experiment.owner_id != current_user.id:
+            raise HTTPException(
+                status_code = 403,
+                detail = "Not authorized"
             )
         experiment_data = experiment.model_dump(exclude_unset = True)
         db_experiment.sqlmodel_update(experiment_data)
@@ -216,6 +272,11 @@ async def delete_experiment(current_user: Annotated[User, Depends(get_current_us
                 status_code = 404,
                 detail = "Experiment not found"
             )
+        if experiment.owner_id != current_user.id:
+            raise HTTPException(
+                status_code = 403,
+                detail = "Not authorized"
+            )
         session.delete(experiment)
         session.commit()
         return {"ok": True}
@@ -224,7 +285,7 @@ async def delete_experiment(current_user: Annotated[User, Depends(get_current_us
 @app.get("/experiments", response_model = list[ExperimentPublic])
 async def get_experiments(current_user: Annotated[User, Depends(get_current_user)]):
     with Session(engine) as session:
-        experiments = session.exec(select(Experiment)).all()
+        experiments = session.exec(select(Experiment).where(Experiment.owner_id == current_user.id)).all()
         return experiments
     
 @app.get("/experiments/{experiment_id}", response_model = ExperimentPublic)
@@ -235,6 +296,11 @@ async def get_experiment(current_user: Annotated[User, Depends(get_current_user)
             raise HTTPException(
                 status_code = 404,
                 detail = "Experiment not found"
+            )
+        if experiment.owner_id != current_user.id:
+            raise HTTPException(
+                status_code = 403,
+                detail = "Not authorized"
             )
         return experiment
